requirements.txt
numpy
pandas
yfinance
matplotlib
google-generativeai
newsapi-python
scikit-learn
python-dotenv
streamlit
openai
plotly
seaborn